fix: apply open upstream issues/PRs — EMCY bit, gateway buffer, warnings
UPSTREAM PR #590 — CO_Emergency.h: CO_EM_RPDO_WRONG_LENGTH moved from
0x04 (communication info) to 0x10 (communication critical).
CiA 301 compliance tests require RPDO wrong-length to be a critical
error that sets the communication bit in the error register, not just
an informational event. Old bit 0x04 is now CO_EM_04_unused.

UPSTREAM ISSUE #623 — CO_gateway_ascii.c: tok[20] -> tok[21].
The command string "lss_activate_bitrate" is exactly 20 characters.
CO_fifo_readToken() with count=20 reads up to 20 chars; when exactly
20 chars are read tokenSize==count which triggers the overflow guard
(*err=true, token discarded). The buffer needs at least 21 bytes
(20 chars + null terminator) to parse this command successfully.

UPSTREAM PR #617 — compiler warning fixes across core 301/ files:
- 301/CO_PDO.c: PDO_initMapping's pdoDataLength: size_t -> uint8_t.
  Max value is 64 bits or 8 bytes, both fit in uint8_t. Fixes
  -Wconversion truncation warning. Added #include <stdint.h>.
- 301/CO_PDO.h: Added #include <stdint.h>.
- 301/CO_ODinterface.c: Explicit (uint16_t) casts in OD_find() binary
  search: od->size-1U, (min+max)>>1, cur-1U. Fixes -Wconversion.
  Added #include <stdint.h>.
- 301/CO_ODinterface.h: Added #include <stdint.h>. Explicit (uint8_t)
  cast in OD_requestTPDO() bitmask negation. Fixes -Wconversion.
- 301/CO_HBconsumer.c: Explicit (uint8_t) casts in subIndex-1U and
  subEntriesCount-1U. Fixes -Wconversion. Added #include <stdint.h>.
- 301/CO_HBconsumer.h: Added #include <stdint.h>.

NOTE: upstream PR #580 (CRC rename crc16->CO_crc16) was already
applied in our fork (301/CO_crc16-ccitt.h/c + Kconfig docs fix
in BUG-R6-002).

+++ 301/CO_ODinterface.c
@@ -17,6 +17,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
+#include <stdint.h>
 #include <string.h>
 #define OD_DEFINITION
 #include "301/CO_ODinterface.h"
@@ -143,13 +144,13 @@ OD_find(OD_t* od, uint16_t index) {
     }
 
     uint16_t min = 0;
-    uint16_t max = od->size - 1U;
+    uint16_t max = od->size - (uint16_t)1U; /* PR #617: explicit cast avoids -Wconversion */
 
     /* Fast search in ordered Object Dictionary. If indexes are mixed, this won't work. If Object
      * Dictionary has up to N entries, then the max number of loop passes is log2(N) */
     while (min < max) {
         /* get entry between min and max */
-        uint16_t cur = (min + max) >> 1;
+        uint16_t cur = (uint16_t)(min + max) >> 1; /* PR #617: explicit cast */
         OD_entry_t* entry = &od->list[cur];
 
         if (index == entry->index) {
@@ -157,7 +158,7 @@ OD_find(OD_t* od, uint16_t index) {
         }
 
         if (index < entry->index) {
-            max = (cur > 0U) ? (cur - 1U) : cur;
+            max = (cur > 0U) ? (cur - (uint16_t)1U) : cur; /* PR #617: explicit cast */
         } else {
             min = cur + 1U;
         }
